update 3h47 8/3

+++ Demo_4/Inc/user_LCD_object.h
@@ -4,11 +4,11 @@
 #include "LCD_object.h"
 #include "user_LCD.h"
 
-void LCD_Change_State_Setup_T1_T2_T3(uint16_t stampTime1, uint16_t stampTime2, uint16_t stampTime3);
+void LCD_Change_State_Setup_T1_T2_T3(uint32_t stampTime1, uint32_t stampTime2, uint32_t stampTime3);
 void LCD_Change_State_Time_HH_MM_SS(uint16_t hh, uint16_t mm, uint16_t ss);
 
 void UintTime_To_CharTime_HH_MM_SS(uint16_t hh, uint16_t mm, uint16_t ss);
-void UintTime_To_CharTime_T1_T2_T3(uint16_t stampTime1, uint16_t stampTime2, uint16_t stampTime3);
+void UintTime_To_CharTime_T1_T2_T3(uint32_t stampTime1, uint32_t stampTime2, uint32_t stampTime3);
 
 
 void USER_LCD_Display_Time(CLCD_Name* LCD);
